requirements.txt
PyGithub
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
jira
python-dotenv
pytest